requirements.txt
chroma-hnswlib==0.7.6
chromadb==0.5.23
google-ai-generativelanguage==0.6.18
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
googleapis-common-protos==1.72.0
langchain==0.3.0
langchain-chroma==0.1.4
langchain-community==0.3.0
langchain-core==0.3.63
langchain-google-genai==2.1.5
langchain-ollama
langchain-huggingface==0.1.0
langchain-text-splitters==0.3.8
unstructured==0.18.21
unstructured-client==0.42.4
fastapi
uvicorn
requests
python-dotenv
msgraph-core
azure-identity
pypdf
sentence-transformers
pdf2image
pdfminer.six
pillow
docx2txt
networkx
python-docx
